integer casting not taking place automatically with / and % operations in python 3.5. Corrected examples to enforce the integer casting on array indices

+++ examples/plot_bids_analysis.py
@@ -84,7 +84,7 @@ for midx, (model, imgs, events, confounds) in enumerate(model_and_args):
     zmap = model.compute_contrast('language-string')
     plotting.plot_glass_brain(zmap, colorbar=False, threshold=norm.isf(0.001),
                               title=('sub-' + model.subject_label),
-                              axes=axes[midx / 5, midx % 5],
+                              axes=axes[int(midx / 5), int(midx % 5)],
                               plot_abs=False, display_mode='x')
 fig.suptitle('subjects z_map language network (language - string)')
 plt.show()
